add label for CI build URL to refer back to circleci
This allows to inspect a built image and refer back to
a permanent link for the build on circleci.

It depends on env var `CIRCLE_BUILD_URL`, which gets set
on every build as described at
https://circleci.com/docs/environment-variables/

The Makefile gets executed on circleci on merge to master,
so the env var automatically exists for the build-and-publish
steps within cricleci.